requirements.txt
streamlit>=1.28.0,<3
pandas>=2.0.0,<3
openpyxl>=3.1.0
python-dotenv>=1.0.0
